Fix: Route for Student Dashboard
Shifted the Dashboard page from the home directory to it's very own page.
